docs: modifies workflows examples and docs
