+++ python/docs/workflows.md
@@ -101,9 +101,9 @@ _Source: [examples/workflows/simple.py](/python/examples/workflows/simple.py)_
 
 ### Multi-Step Workflow
 
-This advanced example showcases a workflow that implements multiplication through repeated addition—demonstrating control flow, state manipulation, and conditional logic.
+This advanced example showcases a workflow that implements multiplication through repeated addition—demonstrating control flow, state manipulation, nesting, and conditional logic.
 
-<!-- embedme examples/workflows/advanced.py -->
+<!-- embedme examples/workflows/nesting.py -->
 
 ```py
 import asyncio
@@ -171,7 +171,7 @@ if __name__ == "__main__":
 
 ```
 
-_Source: [examples/workflows/advanced.py](/python/examples/workflows/advanced.py)_
+_Source: [examples/workflows/nesting.py](/python/examples/workflows/nesting.py)_
 
 This workflow demonstrates several powerful concepts:
 - Implementing loops by returning `Workflow.SELF`
@@ -191,7 +191,70 @@ Workflow nesting allows complex behaviors to be encapsulated as reusable compone
 <!-- embedme examples/workflows/nesting.py -->
 
 ```py
-# Coming soon
+import asyncio
+import sys
+import traceback
+from typing import Literal, TypeAlias
+
+from pydantic import BaseModel
+
+from beeai_framework.errors import FrameworkError
+from beeai_framework.workflows.workflow import Workflow, WorkflowReservedStepName
+
+WorkflowStep: TypeAlias = Literal["pre_process", "add_loop", "post_process"]
+
+
+async def main() -> None:
+    # State
+    class State(BaseModel):
+        x: int
+        y: int
+        abs_repetitions: int | None = None
+        result: int | None = None
+
+    def pre_process(state: State) -> WorkflowStep:
+        print("pre_process")
+        state.abs_repetitions = abs(state.y)
+        return "add_loop"
+
+    def add_loop(state: State) -> WorkflowStep | WorkflowReservedStepName:
+        if state.abs_repetitions and state.abs_repetitions > 0:
+            result = (state.result if state.result is not None else 0) + state.x
+            abs_repetitions = (state.abs_repetitions if state.abs_repetitions is not None else 0) - 1
+            print(f"add_loop: intermediate result {result}")
+            state.abs_repetitions = abs_repetitions
+            state.result = result
+            return Workflow.SELF
+        else:
+            return "post_process"
+
+    def post_process(state: State) -> WorkflowReservedStepName:
+        print("post_process")
+        if state.y < 0:
+            result = -(state.result if state.result is not None else 0)
+            state.result = result
+        return Workflow.END
+
+    multiplication_workflow = Workflow[State, WorkflowStep](name="MultiplicationWorkflow", schema=State)
+    multiplication_workflow.add_step("pre_process", pre_process)
+    multiplication_workflow.add_step("add_loop", add_loop)
+    multiplication_workflow.add_step("post_process", post_process)
+
+    response = await multiplication_workflow.run(State(x=8, y=5))
+    print(f"result: {response.state.result}")
+
+    response = await multiplication_workflow.run(State(x=8, y=-5))
+    print(f"result: {response.state.result}")
+
+
+if __name__ == "__main__":
+    try:
+        asyncio.run(main())
+    except FrameworkError as e:
+        traceback.print_exc()
+        sys.exit(e.explain())
+
+
 ```
 
 _Source: [examples/workflows/nesting.py](/python/examples/workflows/nesting.py)_
@@ -207,7 +270,7 @@ import asyncio
 import sys
 import traceback
 
-from beeai_framework.agents.bee.agent import AgentExecutionConfig
+from beeai_framework.agents.types import AgentExecutionConfig
 from beeai_framework.backend.chat import ChatModel
 from beeai_framework.backend.message import UserMessage
 from beeai_framework.errors import FrameworkError
@@ -344,7 +407,7 @@ This pattern demonstrates:
 ## Examples
 
 - [simple.py](/python/examples/workflows/simple.py) - Basic workflow example
-- [advanced.py](/python/examples/workflows/advanced.py) - More complex workflow with loops
+- [nesting.py](/python/examples/workflows/nesting.py) - More complex workflow with loops
 - [memory.py](/python/examples/workflows/memory.py) - Using memory in workflows
 - [multi_agents.py](/python/examples/workflows/multi_agents.py) - Multi-agent workflow
 - [workflows.ipynb](/python/examples/notebooks/workflows.ipynb) - Interactive notebook examples
