+++ python/examples/workflows/searx_agent.py
